Refactoring: чистка интерфейса IEdge
