+++ GraphLight.Core/Graph/EdgeData.cs
@@ -1,4 +1,6 @@
-﻿namespace GraphLight.Graph
+﻿using System.Drawing;
+
+namespace GraphLight.Graph
 {
     public class EdgeData : BaseViewModel, IEdgeData
     {
@@ -6,6 +8,14 @@
         private bool _isHighlighted;
         private bool _isSelected;
         private int _zIndex;
+        private double _thickness;
+        private string _color;
+
+        public EdgeData()
+        {
+            Color = "Black";
+            Thickness = 1;
+        }
 
         public bool IsSelected
         {
@@ -35,5 +45,22 @@
             set => SetProperty(ref _category, value);
         }
 
+        public string Color
+        {
+            get => _color;
+            set
+            {
+                SetProperty(ref _color, value);
+                RaisePropertyChanged(nameof(StrokeBrush));
+            }
+        }
+
+        public double Thickness
+        {
+            get => _thickness;
+            set => SetProperty(ref _thickness, value);
+        }
+
+        public string StrokeBrush => Color;
     }
 }
