+++ GraphLight.Core/Graph/IEdge.cs
@@ -10,10 +10,6 @@ namespace GraphLight.Graph
         IList<Point2D> PolygonPoints { get; set; }
         IList<Point2D> Points { get; set; }
         IList<Point2D> DraggablePoints { get; set; }
-        double Length { get; }
-        string Color { get; set; }
-        double Thickness { get; set; }
-        void Revert();
         void UpdatePoint(Point2D data);
         void FixDraggablePoints(Point2D data);
         void UpdateSrcPort();
