+++ GraphLight.Core/Graph/IEdgeData.cs
@@ -6,5 +6,7 @@
         bool IsHighlighted { get; set; }
         int ZIndex { get; set; }
         string Category { get; set; }
+        string Color { get; set; }
+        double Thickness { get; set; }
     }
 }
